+++ HISTORY.rst
@@ -5,8 +5,6 @@ History
 TODO
 ^^^^
 
-* Define phenotype attribute as a ``mongoengine.DynamicDocument`` field
-* Import phenotypes (from a *source* dataset to a *destination* dataset)
 * Deal with multi-sheets ``.xlsx`` documents
 * Check chromosomes in *Variants locations*: mind to **scaffold**, **null**, and
   **non-autosomal** chromosomes for *Goat* and *Sheep*
@@ -17,6 +15,8 @@ TODO
 Features
 ^^^^^^^^
 
+* Import phenotypes (from a *source* dataset to a *destination* dataset)
+* Define phenotype attribute as a ``mongoengine.DynamicDocument`` field
 * Import metadata or phenotype *by breeds* or *by samples*
 * Import metadata (from a *source* dataset to a *destination* dataset)
 * Forcing ``plink`` **chrom** options when converting in binary formats